[backend] organizes fixtures
