+++ opencti-platform/opencti-graphql/tests/01-unit/database/stix-2-0-converter-fixtures/SCOs/mac-addr.ts
